cloud_functions/bq-stage-changes/requirements.txt
absl-py>=0.9.0
google-cloud-bigquery>=2.18.0
google-cloud-storage>=1.38.0
google-cloud-tasks>=2.3.0
